seal BitSet.
